Feat/팔로우, 팔로워, 프로필 기능 추가 2025.11.11~2025.11.14

+++ src/main/java/com/teamgreen/makeplan/server/repository/UserRepository.java
@@ -1,15 +1,13 @@
 package com.teamgreen.makeplan.server.repository;
 
 import com.teamgreen.makeplan.server.entity.User;
-import com.teamgreen.makeplan.server.entity.user.UserDocument;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     Boolean existsByEmail(String email);
     Optional<User> findByEmail(String email);
 }
